Fix the displaying of mic devices in Capture settings
The fix was done by caching the mic devices in preload.html to SettingsJson under the variable name "micDevicesCache".

Due to the current limitations of WebView2, you cannot request a list of user devices if the page is not secure (https), since the app is loaded locally, there is no https. However, it is secure, only in "preload.html".