Anti-cheat-safe fullpath lookup
By being able to access a process's fullpath, we can parse the path to assume that a non-whitelist process is in fact a game.

This removes the need to access ProcessModuleCollection which is sometimes denied access by anti-cheat services. ProcessModuleCollection was used to check to see if graphic dll/modules were in a process and assumed it was a game.