feat: AutoDetectGame by GPU usage
AutoDetectGame will check if an unknown process is using over 10% GPU usage in order to determine if it is a game.

There could be improvements, such as waiting a few seconds for graphics to fully load, maybe even factor in CPU usage.

This mainly solves most false positives, which are more annoying and wastes processing time.